requirements.txt
torch==2.0.0
torchvision==0.15.1
numpy==1.23.5
scipy==1.11.4
pandas==1.5.3
scikit-learn==1.2.2
anndata==0.8.0
scanpy==1.9.3
squidpy==1.2.2
networkx==3.1
numba==0.56.4
llvmlite==0.39.1
dask-image==2022.9.0
matplotlib==3.8.2
plotnine==0.12.4
pyyaml>=6.0
tqdm>=4.66.0

